Rename chat-agents to chat-agents-executor
