+++ microservices/chat-agents-executor/chat_agents_executor/singleton.py
